Political_Analyst_Workbench/requirements.txt
langgraph
langchain_openai
langchain_community
tavily-python
python-dotenv
httpx
streamlit
langfuse
boto3
